[SPARK-43919][SQL] Extract JSON functionality out of Row
### What changes were proposed in this pull request?

Extract JSON functionality out of Row. Row is public API that is needed by Spark Connect client. We are planning to move Row to a shared module. However Json functionality is unstable API that is not needed by the client. So we can leave that within catalyst module.

### Why are the changes needed?

Towards Spark Connect client not depend on catalyst and core.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

Closes #41425 from amaliujia/row-json.

Lead-authored-by: Rui Wang <rui.wang@databricks.com>
